:sparkles: feat: ajusta o tipo do campo valor para receber numeros float e salvar desta forma no banco e quando para o pargarme transformar o valor na zod multiplicando por 100

+++ src/backend-worker-financeiro/controllers/conta_receber.ts
@@ -72,7 +72,7 @@ namespace ControllerContaReceber {
                         checkout: CheckoutSchema,
                         cliente_id: z4.string(),
                         parcelas: z4.number(),
-                        valor: z4.number().int(),
+                        valor: z4.number(),
                         vencimento: z4.iso.datetime().optional(),
                         codigo: z4.string(),
                         metodo_pagamento: z4.array(MetodoPagamentoSchema).min(1),
